refactor: tighten review phase and error seams
