+++ tests/orchestration/test_phases.py
@@ -3,6 +3,7 @@
 from __future__ import annotations
 
 import asyncio
+from datetime import UTC, datetime
 from fractions import Fraction
 from pathlib import Path
 
@@ -13,7 +14,9 @@ from frame_compare.orchestration.context import (
     ClipState,
     RunContext,
 )
+from frame_compare.orchestration.execution import build_phases_after_align
 from frame_compare.orchestration.phases import Phase, PhaseStatus, execute_phases
+from frame_compare.orchestration.types import RunArtifacts, RunRequest
 from frame_compare.utils.progress import NullProgressReporter
 from frame_compare.utils.types import WorkspacePaths
 
@@ -267,3 +270,28 @@ def test_execute_phases_empty_list_noop(tmp_path: Path) -> None:
     reporter = NullProgressReporter()
 
     asyncio.run(execute_phases([], context, reporter))
+
+
+def test_publish_phase_skip_condition_uses_effective_slowpics_config() -> None:
+    phase_timings: dict[str, float] = {}
+    warnings: list[str] = []
+    artifacts = RunArtifacts()
+
+    phases = build_phases_after_align(
+        request=RunRequest(root=Path("."), no_upload=False),
+        clock=lambda: datetime(2026, 5, 21, tzinfo=UTC),
+        ffmpeg_runner=object(),
+        http_client=None,
+        phase_timings=phase_timings,
+        warnings=warnings,
+        selected_frames=[],
+        artifacts=artifacts,
+        metadata_prefetched=False,
+    )
+
+    publish_phase = next(phase for phase in phases if phase.name == "publish")
+    config = ConfigSchema()
+    config.slowpics.auto_upload = False
+
+    assert publish_phase.skip_condition is not None
+    assert publish_phase.skip_condition(config) is True
